inotify: rework architecture
Make several changes to the file system change monitoring architecture
to enable further improvements over time.  First, rename the
cola.inotify module to cola.fsmonitor, because inotify is a
Linux-specific implementation detail.  Second, instead of having a
single monitor thread class (GitNotifier) that called an alternate run()
method on Windows, make two separate monitor thread classes
(_InotifyThread and _Win32Thread) that both inherit some common
functionality from the new _BaseThread class.  Third, change the
notification mechanism to use Qt's signal/slot mechanism instead of a
custom list of observer functions.  This allows some code removal in the
cola.app.ColaApplication class which took care of making sure the actual
processing happened in the main thread, because Qt's signal/slot
mechanism handles that automatically.  Finally, handle all interaction
with the fsmonitor module through a singleton instance of the _Monitor
class, accessed through the cola.fsmonitor.instance() function.
