+++ cola/app.py
@@ -62,10 +62,10 @@ from PyQt4.QtCore import SIGNAL
 from cola import cmds
 from cola import core
 from cola import compat
+from cola import fsmonitor
 from cola import git
 from cola import gitcfg
 from cola import icons
-from cola import inotify
 from cola import i18n
 from cola import qtcompat
 from cola import qtutils
@@ -170,11 +170,8 @@ class ColaApplication(object):
         qtutils.install()
         icons.install()
 
-        self.notifier = QtCore.QObject()
-        self.notifier.connect(self.notifier, SIGNAL('update_files()'),
-                              self._update_files, Qt.QueuedConnection)
-        # Call _update_files when inotify detects changes
-        inotify.observer(self._update_files_notifier)
+        QtCore.QObject.connect(fsmonitor.instance(), SIGNAL('files_changed'),
+                               self._update_files)
 
         if gui:
             self._app = current(tuple(argv))
@@ -198,11 +195,8 @@ class ColaApplication(object):
             self._app.view = view
 
     def _update_files(self):
-        # Respond to inotify updates
+        # Respond to file system updates
         cmds.do(cmds.Refresh)
-
-    def _update_files_notifier(self):
-        self.notifier.emit(SIGNAL('update_files()'))
 
 
 @memoize
@@ -300,8 +294,8 @@ def application_start(context, view):
     runtask = qtutils.RunTask(parent=view)
     init_update_task(view, runtask, context.model)
 
-    # Start the inotify thread
-    inotify.start()
+    # Start the filesystem monitor thread
+    fsmonitor.instance().start()
 
     msg_timer = QtCore.QTimer()
     msg_timer.setSingleShot(True)
@@ -312,7 +306,7 @@ def application_start(context, view):
     result = context.app.exec_()
 
     # All done, cleanup
-    inotify.stop()
+    fsmonitor.instance().stop()
     QtCore.QThreadPool.globalInstance().waitForDone()
 
     tmpdir = utils.tmpdir()
